requirements.txt
streamlit>=1.29.0
supabase>=2.0.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
requests>=2.31.0
pydantic>=2.4.0
